requirements.txt
python-telegram-bot==20.7
openai==1.30.1
httpx>=0.25.0,<0.26.0
python-dotenv
